Refactor redux codes
