+++ src/_actions/types.js
@@ -1,0 +1,2 @@
+export const LOGIN_USER = "login_user";
+export const LOGOUT_USER = "logout_user";
