+++ src/_reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from "redux";
+import user from "./user";
+
+const rootReducer = combineReducers({ user });
+
+export default rootReducer; // used as reducer in _app.js
