+++ src/components/molecules/Button/LogIn.js
@@ -1,21 +1,17 @@
 import { useCallback } from "react";
 import GoogleLogin from "react-google-login";
 import { useDispatch } from "react-redux";
+import { loginUser } from "../../../_actions/user";
 
 export default function SignInButton() {
   const dispatch = useDispatch();
   const loginHandler = useCallback((response) => {
-    const name = response.profileObj.name;
-    const email = response.profileObj.email;
-    const image = response.profileObj.imageUrl;
-
-    dispatch({
-      type: "IN",
-      isSignedIn: true,
-      name: name,
-      email: email,
-      profile_pic: image,
-    });
+    let userData = {
+      name: response.profileObj.name,
+      email: response.profileObj.email,
+      profilePic: response.profileObj.imageUrl,
+    };
+    dispatch(loginUser(userData));
   }, []);
 
   return (
